+++ src/lib.rs
@@ -92,22 +92,8 @@ use std::path::{Path, PathBuf};
 use std::process::Command;
 use std::{env, fs};
 
-pub mod ffi {
-    #![allow(
-        improper_ctypes,
-        non_upper_case_globals,
-        non_camel_case_types,
-        non_snake_case
-    )]
-    #![allow(
-        clippy::pedantic,
-        clippy::approx_constant,
-        clippy::useless_transmute,
-        clippy::too_many_arguments
-    )]
-    include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
-}
-
+// Re-publish varnish_sys::ffi
+pub use varnish_sys::ffi;
 pub mod vsc;
 
 mod error;
