+++ varnish-sys/README.md
@@ -1,0 +1,3 @@
+# varnish-sys
+
+Some unsafe and auto-generated bindings for the Varnish Cache project.
